// frontend-v2/src/api/generated.ts
export interface components {
    schemas: {
// ...
        JobOut: {
            /**
             * Id
             * Format: uuid
             */
            id: string;
            /** Book Id */
            book_id: string | null;
            /** Type */
            type: string;
            /** Status */
            status: string;
            /** Progress */
            progress: number;
            /** Message */
            message: string | null;
            /** Error */
            error: string | null;
            /** Started At */
            started_at: string | null;
            /** Finished At */
            finished_at: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}